+++ mojo/edk/system/core_test_base.cc
@@ -172,7 +172,7 @@ class MockDispatcher : public Dispatcher {
 
   MojoResult AddAwakableImplNoLock(Awakable* awakable,
                                    MojoHandleSignals /*signals*/,
-                                   uint32_t /*context*/,
+                                   uint64_t /*context*/,
                                    HandleSignalsState* signals_state) override {
     info_->IncrementAddAwakableCallCount();
     mutex().AssertHeld();
