+++ mojo/edk/system/data_pipe_impl_unittest.cc
@@ -141,7 +141,7 @@ class DataPipeImplTest : public testing::Test {
   }
   MojoResult ProducerAddAwakable(Awakable* awakable,
                                  MojoHandleSignals signals,
-                                 uint32_t context,
+                                 uint64_t context,
                                  HandleSignalsState* signals_state) {
     return dpp()->ProducerAddAwakable(awakable, signals, context,
                                       signals_state);
@@ -180,7 +180,7 @@ class DataPipeImplTest : public testing::Test {
   }
   MojoResult ConsumerAddAwakable(Awakable* awakable,
                                  MojoHandleSignals signals,
-                                 uint32_t context,
+                                 uint64_t context,
                                  HandleSignalsState* signals_state) {
     return dpc()->ConsumerAddAwakable(awakable, signals, context,
                                       signals_state);
@@ -292,7 +292,7 @@ class RemoteDataPipeImplTestHelper : public DataPipeImplTestHelper {
                                     MOJO_WRITE_MESSAGE_FLAG_NONE));
       transport.End();
     }
-    uint32_t context = 0;
+    uint64_t context = 0;
     ASSERT_EQ(MOJO_RESULT_OK, waiter.Wait(test::ActionTimeout(), &context));
     EXPECT_EQ(987u, context);
     HandleSignalsState hss = HandleSignalsState();
@@ -638,7 +638,7 @@ TYPED_TEST(DataPipeImplTest, SimpleReadWrite) {
 
   Waiter waiter;
   HandleSignalsState hss;
-  uint32_t context;
+  uint64_t context;
 
   int32_t elements[10] = {};
   uint32_t num_bytes = 0u;
@@ -792,7 +792,7 @@ TYPED_TEST(DataPipeImplTest, BasicProducerWaiting) {
   Waiter pwaiter;  // For producer.
   Waiter cwaiter;  // For consumer.
   HandleSignalsState hss;
-  uint32_t context;
+  uint64_t context;
 
   // Never readable.
   pwaiter.Init();
@@ -995,7 +995,7 @@ TYPED_TEST(DataPipeImplTest, PeerClosedProducerWaiting) {
 
   Waiter waiter;
   HandleSignalsState hss;
-  uint32_t context;
+  uint64_t context;
 
   // Add a waiter.
   waiter.Init();
@@ -1030,7 +1030,7 @@ TYPED_TEST(DataPipeImplTest, PeerClosedConsumerWaiting) {
 
   Waiter waiter;
   HandleSignalsState hss;
-  uint32_t context;
+  uint64_t context;
 
   // Add a waiter.
   waiter.Init();
@@ -1066,7 +1066,7 @@ TYPED_TEST(DataPipeImplTest, BasicConsumerWaiting) {
   Waiter waiter;
   Waiter waiter2;
   HandleSignalsState hss;
-  uint32_t context;
+  uint64_t context;
 
   // Never writable.
   waiter.Init();
@@ -1255,7 +1255,7 @@ TYPED_TEST(DataPipeImplTest, ConsumerWaitingTwoPhase) {
 
   Waiter waiter;
   HandleSignalsState hss;
-  uint32_t context;
+  uint64_t context;
 
   // Add waiter: not yet readable.
   waiter.Init();
